+++ packages/core/src/components/MultiSelect/Options/Chip/Chip.test.tsx
@@ -5,14 +5,14 @@ import { Chip } from './Chip';
 describe('Chip of MultiSelect component', () => {
     it('should call onClear on click of cross icon', () => {
         const onClearMock = jest.fn(),
-            { container } = render(<Chip size="S" label={2} onClear={onClearMock} />);
+            { container } = render(<Chip label={2} onClear={onClearMock} />);
         fireEvent.click(container.querySelector('svg'));
         expect(onClearMock).toHaveBeenCalledWith(2);
     });
 
     it('should not call onClear if the chip is disabled', () => {
         const onClearMock = jest.fn(),
-            { container } = render(<Chip size="M" label={2} disabled={true} onClear={onClearMock} />);
+            { container } = render(<Chip label={2} disabled={true} onClear={onClearMock} />);
         fireEvent.click(container.querySelector('svg'));
         expect(onClearMock).not.toHaveBeenCalled();
     });