+++ packages/utils/src/helpers/isMobile.ts
@@ -1,0 +1,1 @@
+export const isMobile = /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(navigator.userAgent);