+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.10](https://github.com/medly/medly-components/compare/@medly-components/icons@1.4.9...@medly-components/icons@1.4.10) (2020-11-10)
+
+**Note:** Version bump only for package @medly-components/icons
+
+
+
+
+
+## [1.4.9](https://github.com/medly/medly-components/compare/@medly-components/icons@1.4.8...@medly-components/icons@1.4.9) (2020-11-05)
+
+**Note:** Version bump only for package @medly-components/icons
+
+
+
+
+
+## [1.4.8](https://github.com/medly/medly-components/compare/@medly-components/icons@1.4.7...@medly-components/icons@1.4.8) (2020-10-30)
+
+**Note:** Version bump only for package @medly-components/icons
+
+
+
+
+
 ## [1.4.7](https://github.com/medly/medly-components/compare/@medly-components/icons@1.4.6...@medly-components/icons@1.4.7) (2020-10-28)
 
 **Note:** Version bump only for package @medly-components/icons