+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.3](https://github.com/medly/medly-components/compare/@medly-components/utils@1.6.2...@medly-components/utils@1.6.3) (2020-11-10)
+
+**Note:** Version bump only for package @medly-components/utils
+
+
+
+
+
+## [1.6.2](https://github.com/medly/medly-components/compare/@medly-components/utils@1.6.1...@medly-components/utils@1.6.2) (2020-11-05)
+
+**Note:** Version bump only for package @medly-components/utils
+
+
+
+
+
+## [1.6.1](https://github.com/medly/medly-components/compare/@medly-components/utils@1.6.0...@medly-components/utils@1.6.1) (2020-10-30)
+
+**Note:** Version bump only for package @medly-components/utils
+
+
+
+
+
 # [1.6.0](https://github.com/medly/medly-components/compare/@medly-components/utils@1.5.5...@medly-components/utils@1.6.0) (2020-10-28)